fix logging in serf
before I was sending all serf log to syslog which then they all go
to the fronend, so with big cluster that ain't no good. Moreover
there is a bug (which is fixed in the mainline) which send all
logs included debug to syslog.

So now logs are local to each node in /var/log/serf.log the file
is rotated under logrotate